requirements.txt
asgiref==3.7.2
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==3
pytz==2024.1
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
